added forwared

+++ car_controller.py
@@ -20,7 +20,7 @@ def forward(speed=0.5, time=0):
     if time < 0:
         raise ValueError("Can't wait negativ Time")
 
-    for speed_tmp in [x * 0.1 for x in range(0, (speed*10))]:
+    for speed_tmp in [x * 0.1 for x in range(0, int(speed*10))]:
         base.set_speed(speed_tmp, speed_tmp)
         time.sleep(0.2)
